+++ Climate/Climate_Plugin.cs
@@ -1,6 +1,7 @@
 ﻿using BepInEx;
 using BepInEx.Logging;
 using HarmonyLib;
+using System.Collections.Generic;
 using System.Reflection;
 using UnityEngine.SceneManagement;
 
@@ -11,7 +12,7 @@ namespace Climate
     {
         public const string PLUGIN_GUID = "com.raddude.climate";
         public const string PLUGIN_NAME = "Climate";
-        public const string PLUGIN_VERSION = "1.0.1";
+        public const string PLUGIN_VERSION = "1.1.0";
 
         internal static Climate_Plugin Instance { get; private set; }
         private static ManualLogSource _logger;
@@ -20,6 +21,8 @@ namespace Climate
         internal static void LogInfo(string message) => _logger.LogInfo(message);
         internal static void LogWarning(string message) => _logger.LogWarning(message);
         internal static void LogError(string message) => _logger.LogError(message);
+
+        internal static List<PressureCell> PressureCells => PressureService.cells;
 
         private void Awake()
         {
@@ -37,6 +40,8 @@ namespace Climate
 
             Harmony.CreateAndPatchAll(Assembly.GetExecutingAssembly(), PLUGIN_GUID);
             SceneManager.sceneLoaded += AddShopItems.SceneLoaded;
-        }
+
+            Sun.OnNewDay += PressureService.UpdatePressureCells;
+        }        
     }
 }