+++ Climate/Patches/SaveLoadPatches.cs
@@ -1,0 +1,25 @@
+﻿using HarmonyLib;
+
+namespace Climate.Patches
+{
+    internal class SaveLoadPatches
+    {
+        [HarmonyPatch(typeof(SaveLoadManager))]
+        private class SaveLoadManagerPatches
+        {
+            [HarmonyPrefix]
+            [HarmonyPatch("SaveModData")]
+            public static void SaveModData()
+            {
+                PressureService.SavePressureCells();
+            }
+
+            [HarmonyPrefix]
+            [HarmonyPatch("LoadModData")]
+            public static void LoadModData()
+            {
+                PressureService.LoadPressureCells();
+            }
+        }
+    }
+}